feat: improve the UX of HeroSection
