+++ app/native-components/HomeHeroSection/HomeHeroSection.tsx
@@ -1,10 +1,46 @@
 "use client";
 
 import Carousel from "@/components/ui/carousel";
-import React from "react";
+import React, { useState } from "react";
 import { HeroSectionSlides } from "./data";
+import { IconArrowNarrowRight } from "@tabler/icons-react";
+
+interface CarouselControlProps {
+  type: string;
+  title: string;
+  handleClick: () => void;
+}
+
+const CarouselControl = ({
+  type,
+  title,
+  handleClick,
+}: CarouselControlProps) => {
+  return (
+    <button
+      className={`w-10 h-10 flex items-center mx-2 justify-center bg-neutral-200 dark:bg-neutral-800 border-3 border-transparent rounded-full focus:border-[#6D64F7] focus:outline-none hover:-translate-y-0.5 active:translate-y-0.5 transition duration-200 ${
+        type === "previous" ? "rotate-180" : ""
+      }`}
+      title={title}
+      onClick={handleClick}
+    >
+      <IconArrowNarrowRight className="text-neutral-600 dark:text-neutral-200" />
+    </button>
+  );
+};
 
 const HomeHeroSection = () => {
+  const [current, setCurrent] = useState(0);
+
+  const handlePreviousClick = () => {
+    const previous = current - 1;
+    setCurrent(previous < 0 ? HeroSectionSlides.length - 1 : previous);
+  };
+
+  const handleNextClick = () => {
+    const next = current + 1;
+    setCurrent(next === HeroSectionSlides.length ? 0 : next);
+  };
   return (
     <section
       className="relative min-h-screen w-full bg-black flex flex-col items-center justify-center py-20"
@@ -27,8 +63,26 @@ const HomeHeroSection = () => {
           </p>
         </div>
 
-        <div className="max-w-7xl">
-          <Carousel slides={HeroSectionSlides} />
+        <div className="relative overflow-hidden w-full h-full ">
+          <Carousel
+            slides={HeroSectionSlides}
+            current={current}
+            setCurrent={setCurrent}
+          />
+
+          <div className="absolute flex justify-between w-full top-[calc(50%+1rem)] ">
+            <CarouselControl
+              type="previous"
+              title="Go to previous slide"
+              handleClick={handlePreviousClick}
+            />
+
+            <CarouselControl
+              type="next"
+              title="Go to next slide"
+              handleClick={handleNextClick}
+            />
+          </div>
         </div>
       </div>
